added the majority of features of the cmi api and the client actions (create/delete linked with comptePaiement)

+++ src/main/java/com/ensa/jibi/backend/services/impl/ClientServiceImpl.java
@@ -1,13 +1,12 @@
 package com.ensa.jibi.backend.services.impl;
 
 import com.ensa.jibi.backend.domain.dto.ClientDto;
-import com.ensa.jibi.backend.domain.dto.DocumentDto;
 import com.ensa.jibi.backend.domain.entities.Client;
-import com.ensa.jibi.backend.domain.entities.Document;
 import com.ensa.jibi.backend.mappers.ClientMapper;
 import com.ensa.jibi.backend.repositories.ClientRepository;
 import com.ensa.jibi.backend.services.ClientService;
-import org.modelmapper.ModelMapper;
+import com.ensa.jibi.cmi.domain.dto.ComptePaiementDto;
+import com.ensa.jibi.cmi.services.impl.ComptePaiementServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,46 +21,43 @@ public class ClientServiceImpl implements ClientService {
 
     private final ClientRepository clientRepository;
     private final ClientMapper clientMapper;
-    private final ModelMapper modelMapper;
+    private final ComptePaiementServiceImpl comptePaiementService;
 
     @Autowired
-    public ClientServiceImpl(ClientRepository clientRepository, ClientMapper clientMapper, ModelMapper modelMapper) {
+    public ClientServiceImpl(ClientRepository clientRepository, ClientMapper clientMapper, ComptePaiementServiceImpl comptePaiementService) {
         this.clientRepository = clientRepository;
         this.clientMapper = clientMapper;
-        this.modelMapper = modelMapper;
+        this.comptePaiementService = comptePaiementService;
     }
 
     @Override
     public ClientDto addClient(ClientDto clientDto) {
+        String phoneNumber = clientDto.getNumTel();
+        if (comptePaiementService.existsById(phoneNumber)) {
+            throw new IllegalArgumentException("Phone number already exists as ComptePaiement ID");
+        }
+
         Client client = clientMapper.mapFrom(clientDto);
         client = clientRepository.save(client);
+
+        // Create a new ComptePaiement with the ID = phone number
+        ComptePaiementDto comptePaiement = new ComptePaiementDto();
+        comptePaiement.setId(phoneNumber);
+        comptePaiement.setSolde(0.0); // Initialize solde
+        comptePaiementService.save(comptePaiement);
         return clientMapper.mapTo(client);
     }
 
     @Override
     public ClientDto updateClient(Long id, ClientDto clientDto) {
         Optional<Client> existingClientOpt = clientRepository.findById(id);
-        //This line checks if the Optional<Client> object retrieved in the previous step has a value.
         if (existingClientOpt.isPresent()) {
-            //If the client was found (isPresent() returned true), this line retrieves the actual Client object from the Optional.
             Client existingClient = existingClientOpt.get();
 
             // Update basic fields
             existingClient.setEmail(clientDto.getEmail());
             existingClient.setNumTel(clientDto.getNumTel());
             existingClient.setClientType(clientDto.getClientType());
-
-//            // Clear the existing documents and add new ones
-            // Minaoui : disabling update for doc =  they should be static i guess !!
-//            existingClient.getDocuments().clear();
-//            List<Document> documents = clientDto.getDocuments().stream()
-//                    .map(documentDto -> {
-//                        Document document = modelMapper.map(documentDto, Document.class);
-//                        document.setClient(existingClient); // Set the client reference
-//                        return document;
-//                    })
-//                    .collect(Collectors.toList());
-//            existingClient.getDocuments().addAll(documents);
 
             Client updatedClient = clientRepository.save(existingClient);
             return clientMapper.mapTo(updatedClient);
@@ -72,7 +68,17 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public void deleteClient(Long id) {
-        clientRepository.deleteById(id);
+        Optional<Client> clientOpt = clientRepository.findById(id);
+        if (clientOpt.isPresent()) {
+            Client client = clientOpt.get();
+            String phoneNumber = client.getNumTel();
+
+            clientRepository.deleteById(id);
+
+            comptePaiementService.delete(phoneNumber);
+        } else {
+            throw new RuntimeException("Client not found");
+        }
     }
 
     @Override
@@ -88,5 +94,10 @@ public class ClientServiceImpl implements ClientService {
                 .stream()
                 .map(clientMapper::mapTo)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteByNumTel(String id) {
+        clientRepository.deleteByNumTel(id);
     }
 }
